pyproject.toml
[tool.poetry]
name = "cs231n-book"
version = "0.1.0"
description = ""
authors = ["Justin <justinluong1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.41.2"
pypdf = "^4.1.0"
fpdf2 = "^2.7.8"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
types-fpdf2 = "^2.7.8.20240217"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
